feat: add abbybot privileges blueprint and route for privilege management
