+++ abbybot_website/app/__init__.py
@@ -29,6 +29,8 @@ def create_app():
     
     # Register Blueprints
     from app.routes.main import main_bp
+    from app.routes.abbybot_privileges import abbybot_privileges_bp  # Import the new blueprint
     app.register_blueprint(main_bp)
+    app.register_blueprint(abbybot_privileges_bp)  # Register the new blueprint
     
     return app
